requirements.txt
Flask==3.0.3
tensorflow==2.21.0
numpy==2.1.1
opencv-python==4.10.0.84
Pillow==10.4.0
Werkzeug==3.0.4
mtcnn==1.0.0